Added a whole lot of pseudo selectors:
- :first, :last, :empty, :only, :contains, :enabled, :odd, :even

Every unknown pseudoselector will be treated as property. So a:href will be the same as a[href] and a:href(something) will be the same as a[href=something].
Note: everything after and including the minus character (-) will be ignored in a pseudo selector, since we dont plan to support of-type, last-of-type and first-of-type selectors.

git-svn-id: http://svn.mootools.net/trunk@635 4db308c1-fb21-0410-9919-de62e267375e