Minor tweak to typeOf.
Long explanation: HtmlUnit (a headless browser implementation for unit testing) was barfing on typeOf whenever you called it with a string. That's because it evaluated "".$family as "falsy" despite being defined. An explicit check here solved the problem and now *most* of the unit tests for Core pass (not the ones that are fairly browser specific, like element measurements and whatnot).