With the updated specs, all green (related to this commit) for IE6-9 and Chrome latest, Safari latest, Opera, Firefox 3-5.
The fix in Element.js is to overload the setter, getter, and erase to
forcibly update the id attribute (on the object) and with the attribute
node. It's unfortunate, but IE has trouble removing attributes of
cloned nodes.

We should be vigilant of other troublesome attribute nodes that might be
problematic.