Fixes #2251 - Do not overwrite the prototype methods of native types.
It overwrote the methods on the prototype to set enumerable to true, so it was possible to iterate over the methods of the prototypes of types.

Doing this, however, caused performance issues in Chrome: http://code.google.com/p/chromium/issues/detail?id=67772

To still be able to iterate over the methods, the methos array is saved in the $methods property, i.e. Array.$methods.

Methods that are later implemented, with Array.implement are enumerable and will be visible in the for-in loops.