Fixes #2160.
Using @ryanmorr's fix for non px starting values. This fix is primarily
for FFx, but it's not very easy to feature detect this Firefox bug. Open
to suggestions.

For now, this works quite well and also covers em, and other units.

This doesn't work for getStyle since we require an ending value, and
besides we wouldn't want to do this across all units.

PASSES: IE6-9; Firefox 3-5, 8, 10; Opera 11; Chrome latest; Safari 5