Tests: Replace Syn with a local version that works in IE9 and IE10.
Apparently triggering a click on an <a href="javascript:somethin()" />
in IE9 and IE10 causes a page reload of which Karma doesn't approve.

Since we don't use this functionality in our tests anyway, preventing
that trigger from happening has no effects for us.

Diff of the change to the original source (before building):

--- a/src/mouse.support.js
+++ b/src/mouse.support.js
@@ -22,7 +22,7 @@ if (!document.body) {
	select = form.getElementsByTagName('select')[0];

	//trigger click for linkHrefJS support, childNodes[6] === anchor
-	syn.trigger(form.childNodes[6], 'click', {});
+	//syn.trigger(form.childNodes[6], 'click', {});

	checkbox.checked = false;
	checkbox.onchange = function () {