Class.Thenable: Add Promises/A+ tests.
These tests require mocha and will not work in Jasmine 2.3.

The Promises/A+ tests do not run in non-ES5 environments, to make things
easier they're just skipped in all environments where
`require('promises-aplus-tests')` doesn't work. This means that the
Promises/A+ tests are basically only used in server environments (which
should be enough to verify compliance). The "normal" Class.Thenable
specs do run in browsers.